moved deque code to separate file
