+++ MegaInt/CMakeLists.txt
@@ -2,6 +2,7 @@ project(MegaInt LANGUAGES CXX)
 
 set(sources
     src/MegaInt.cpp
+    src/DequeUtils.cpp
 )
 
 add_library(MegaInt ${sources})
