+++ MegaInt/include/MegaInt.h
@@ -4,11 +4,6 @@
 #include <deque>
 #include <string>
 
-namespace
-{
-struct DivisionResult;
-}
-
 /*
 	MegaInt is a class used to describe a data type that models arbitrary large integers.
 	Uses STL deque sequence container to store the digits of the numbers.
@@ -17,7 +12,7 @@ struct DivisionResult;
 class MegaInt
 {
 
-  public:
+public:
 	MegaInt();
 
 	MegaInt(char const new_sign, std::deque<char> &new_magnitude);
@@ -76,40 +71,20 @@ class MegaInt
 
 	friend std::ostream &operator<<(std::ostream &out, const MegaInt &m);
 
-  private:
+private:
 	void Insert(std::string const &new_number);
 
 	void Normalize();
 
 	MegaInt Add(MegaInt B) const;
 
-	// TODO
-	// why are deque based functions part of class def
-	std::deque<char> Plus(std::deque<char> A, std::deque<char> B) const;
-
 	MegaInt Subtract(MegaInt B) const;
-
-	std::deque<char> Minus(std::deque<char> A, std::deque<char> B) const;
 
 	MegaInt Multiply(MegaInt B) const;
 
-	std::deque<char> Product(std::deque<char> A, std::deque<char> B) const;
-
-	std::deque<char> Product2(std::deque<char> A, std::deque<char> B) const;
-
 	MegaInt Division(MegaInt B) const;
 
-	DivisionResult Quotient(std::deque<char> A, std::deque<char> B) const;
-
-	DivisionResult Quotient2(std::deque<char> A, std::deque<char> B) const;
-
 	MegaInt Modulus(MegaInt B) const;
-
-	bool GreaterThan(const std::deque<char> &A, const std::deque<char> &B) const;
-
-	bool LesserThan(const std::deque<char> &A, const std::deque<char> &B) const;
-
-	bool EqualTo(const std::deque<char> &A, const std::deque<char> &B) const;
 
 	char m_sign;
 	std::deque<char> m_magnitude;
